requirements.txt
gspread
PyGithub
pandas
python-dotenv
slack_sdk
oauth2client